requirements.txt
boto3>=1.28.0
pyyaml>=6.0
dataclasses>=0.6
